# contracts/studio-api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Fragments Studio API", "version": "1.0.0", "description": "Authenticated Studio transport contract." },
  "servers": [{ "url": "https://fragments-staging.anchor-event.fr" }],
  "security": [{ "studioOAuth": [] }],
  "paths": {
...
  "/api/admin/experiences/{id}": {
    "get": {
      "operationId": "getAdminExperience",
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Admin snapshot",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminExperienceDetail"
              }
            }
          }
        },
        "401": {
          "description": "Authentication required"
        },
        "403": {
          "description": "Admin access required"
        },
        "404": {
          "description": "Resource not found"
        },
        "400": {
          "description": "Invalid query"
        }
      }
    }
  },
...
  "/api/admin/experience-media/{id}": {
    "get": {
      "operationId": "getAdminExperienceMedia",
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Admin snapshot",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminExperienceMedia"
              }
            }
          }
        },
        "401": {
          "description": "Authentication required"
        },
        "403": {
          "description": "Admin access required"
        },
        "404": {
          "description": "Resource not found"
        },
        "400": {
          "description": "Invalid query"
        }
      }
    }
  },
...
    "/api/admin/media/{id}": {
      "get": {
        "operationId": "getAdminMedia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCatalogItem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Resource not found"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      }
    },
...
    "/api/admin/studio/experience-media/{mediaId}": {
      "get": {
        "operationId": "readExperienceMediaLifecycle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceMediaLifecycle"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Resource not found"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      }
    },
...
},
  "components": {
    "securitySchemes": { "studioOAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
    "schemas": {
...
  "AdminExperience": {
    "type": "object",
    "additionalProperties": false,
    "required": [
      "experienceId",
      "coffeeId",
      "authorId",
      "authorName",
      "avatarUrl",
      "message",
      "publicationStatus",
      "moderationStatus",
      "createdAt",
      "updatedAt",
      "version",
      "media"
    ],
    "properties": {
      "experienceId": {
        "type": "string"
      },
      "coffeeId": {
        "type": "string"
      },
      "authorId": {
        "type": "string"
      },
      "authorName": {
        "type": "string"
      },
      "avatarUrl": {
        "type": [
          "string",
          "null"
        ]
      },
      "message": {
        "type": [
          "string",
          "null"
        ]
      },
      "publicationStatus": {
        "type": "string"
      },
      "moderationStatus": {
        "type": "string"
      },
      "createdAt": {
        "type": "string"
      },
      "updatedAt": {
        "type": "string"
      },
      "version": {
        "type": "integer"
      },
      "media": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ExperienceMedia"
        }
      }
    }
  },
...
  "AdminExperienceMedia": {
    "type": "object",
    "additionalProperties": false,
    "required": [
      "mediaId",
      "experienceId",
      "ownerId",
      "status",
      "url",
      "contentType",
      "updatedAt",
      "size",
      "width",
      "height"
    ],
    "properties": {
      "mediaId": {
        "type": "string"
      },
      "experienceId": {
        "type": "string"
      },
      "ownerId": {
        "type": [
          "string",
          "null"
        ]
      },
      "status": {
        "type": "string"
      },
      "url": {
        "type": [
          "string",
          "null"
        ]
      },
      "contentType": {
        "type": [
          "string",
          "null"
        ]
      },
      "updatedAt": {
        "type": "string"
      },
      "size": {
        "type": "integer"
      },
      "width": {
        "type": [
          "integer",
          "null"
        ]
      },
      "height": {
        "type": [
          "integer",
          "null"
        ]
      }
    }
  },
  "AdminExperienceDetail": {
    "type": "object",
    "additionalProperties": false,
    "required": [
      "experience",
      "media",
      "actions",
      "nextActionsCursor"
    ],
    "properties": {
      "experience": {
        "$ref": "#/components/schemas/AdminExperience"
      },
      "media": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AdminExperienceMedia"
        }
      },
      "actions": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ExperienceModerationAction"
        }
      },
      "nextActionsCursor": {
        "type": [
          "string",
          "null"
        ]
      }
    }
  },
...
      "ExperienceModerationAction": {
        "type": "object", "additionalProperties": false,
        "required": ["actionId", "operatorId", "decision", "occurredAt"],
        "properties": { "actionId": { "type": "string", "format": "uuid" }, "operatorId": { "type": "string", "format": "uuid" }, "decision": { "type": "string", "enum": ["HIDDEN", "VISIBLE"] }, "reason": { "type": ["string", "null"] }, "occurredAt": { "type": "string", "format": "date-time" } }
      },
...
      "ExperienceMedia": {
        "type": "object", "additionalProperties": false,
        "required": ["mediaId", "url", "position"],
        "properties": { "mediaId": { "type": "string", "format": "uuid" }, "url": { "type": "string", "format": "uri" }, "width": { "type": ["integer", "null"] }, "height": { "type": ["integer", "null"] }, "position": { "type": "integer", "minimum": 0 } }
      },
...
      "MediaCatalogItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "mediaId",
          "updatedAt",
          "origin",
          "status",
          "resourceId",
          "ownerId",
          "previewUrl",
          "contentType",
          "createdAt",
          "size",
          "width",
          "height"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "origin": {
            "type": "string",
            "enum": [
              "EXPERIENCE",
              "COFFEE",
              "AVATAR",
              "ARTICLE"
            ]
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "AVAILABLE", "DELETION_PENDING", "DELETED", "REVIEW_REQUIRED", "REJECTED"]
          },
          "resourceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "ownerId": {
            "type": [
              "string",
              "null"
            ]
          },
          "previewUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "contentType": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "size": {
            "type": [
              "integer",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ]
          },
          "height": {
            "type": [
              "integer",
              "null"
            ]
          },
          "originalName": {
            "type": [
              "string",
              "null"
            ]
          },
          "uploadedBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ]
          },
          "usageStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "articleUsages": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MediaCatalogArticleUsagePage"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    "MediaCatalogArticleUsage": {
      "type": "object",
      "required": [
        "articleId",
        "revisionId",
        "revisionNumber",
        "title",
        "revisionStatus",
        "role",
        "sectionPosition",
        "imagePosition",
        "width",
        "height",
        "working",
        "published"
      ],
      "properties": {
        "articleId": {
          "type": "string"
        },
        "revisionId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "revisionStatus": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "revisionNumber": {
          "type": "integer"
        },
        "sectionPosition": {
          "type": [
            "integer",
            "null"
          ]
        },
        "imagePosition": {
          "type": [
            "integer",
            "null"
          ]
        },
        "width": {
          "type": [
            "integer",
            "null"
          ]
        },
        "height": {
          "type": [
            "integer",
            "null"
          ]
        },
        "working": {
          "type": "boolean"
        },
        "published": {
          "type": "boolean"
        }
      }
    },
    "MediaCatalogArticleUsagePage": {
      "type": "object",
      "required": [
        "items",
        "nextCursor"
      ],
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/MediaCatalogArticleUsage"
          }
        },
        "nextCursor": {
          "type": [
            "string",
            "null"
          ]
        }
      }
    },
...
      "ExperienceMediaLifecycle": {
        "type": "object",
        "required": [
          "mediaId",
          "experienceId",
          "coffeeId",
          "ownerId",
          "status",
          "publicationStatus",
          "moderationStatus",
          "canHidePublication",
          "canRestorePublication"
        ],
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "experienceId": {
            "type": "string",
            "format": "uuid"
          },
          "coffeeId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "AVAILABLE", "DELETION_PENDING", "DELETED", "REVIEW_REQUIRED", "REJECTED"]
          },
          "publicationStatus": {
            "type": "string",
            "nullable": true,
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "DELETED"
            ]
          },
          "moderationStatus": {
            "type": "string",
            "nullable": true,
            "enum": [
              "VISIBLE",
              "HIDDEN"
            ]
          },
          "canHidePublication": {
            "type": "boolean"
          },
          "canRestorePublication": {
            "type": "boolean"
          }
        }
      },
...
}
  }
}